requirements.txt
fastapi
uvicorn
motor
sqlalchemy
alembic
passlib[bcrypt]
bcrypt==3.2.0  # Specify a compatible version of bcrypt
python-multipart
python-dotenv
python-jose
google-cloud-firestore
beautifulsoup4==4.12.2
httpx==0.25.2
lxml==4.9.3